@@ -14,21 +14,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.8",
-    "@angular/cdk": "^15.2.9",
-    "@angular/common": "^15.2.8",
-    "@angular/compiler": "^15.2.8",
-    "@angular/core": "^15.2.8",
+    "@angular/animations": "^16.2.12",
+    "@angular/cdk": "^16.2.12",
+    "@angular/cli": "^16.2.12",
+    "@angular/common": "^16.2.12",
+    "@angular/compiler": "^16.2.12",
+    "@angular/core": "^16.2.12",
     "@angular/flex-layout": "^15.0.0-beta.42",
-    "@angular/forms": "^15.2.8",
-    "@angular/material": "^15.2.9",
-    "@angular/platform-browser": "^15.2.8",
-    "@angular/platform-browser-dynamic": "^15.2.8",
-    "@angular/platform-server": "^15.2.8",
-    "@angular/router": "^15.2.8",
+    "@angular/forms": "^16.2.12",
+    "@angular/material": "^16.2.12",
+    "@angular/platform-browser": "^16.2.12",
+    "@angular/platform-browser-dynamic": "^16.2.12",
+    "@angular/platform-server": "^16.2.12",
+    "@angular/router": "^16.2.12",
     "bootstrap": "^5.2.3",
     "jquery": "^3.6.4",
     "ngx-mask": "^17.0.4",
+    "ngx-toastr": "^18.0.0",
     "oidc-client": "^1.11.5",
     "popper.js": "^1.16.0",
     "run-script-os": "^1.1.6",
@@ -37,8 +39,8 @@
     "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.7",
-    "@angular/compiler-cli": "^15.2.8",
+    "@angular-devkit/build-angular": "^16.2.12",
+    "@angular/compiler-cli": "^16.2.12",
     "@types/jasmine": "~4.3.1",
     "@types/jasminewd2": "~2.0.10",
     "@types/node": "^18.16.3",
