@@ -7,7 +7,7 @@ import { Category } from '../../../Models/Category';
 import { Observable } from 'rxjs';
 import { CategoryService } from '../../../Services/CategoryService';
 import { LoadingService } from '../../../Services/LoadingService';
-import { AlertService } from '../../../Services/AlertService';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-create-category',
@@ -24,7 +24,7 @@ export class CreateCategoryComponent implements OnInit {
     private fb: FormBuilder,
     private router: Router,
     private loadingService: LoadingService,
-    private alertService: AlertService
+    private toastr: ToastrService
 
   ) {
     this.category = {
@@ -52,12 +52,12 @@ export class CreateCategoryComponent implements OnInit {
       if (this.categoryForm.valid) {
         const formData: Category = this.categoryForm.value;
         const createdCategory = await (await this.categoryService.createCategory(formData)).toPromise();
-        this.alertService.success(`Categoria ${formData.name} criada com sucesso.`);
+        this.toastr.success(`Categoria ${formData.name} criada com sucesso.`);
         this.router.navigate(['/categories']);
 
       }
     } catch (error: any) {
-      this.alertService.error(error.message || 'Erro interno da aplicação, tente novamente.');
+      this.toastr.error(error.message || 'Erro interno da aplicação, tente novamente.');
       console.error('Erro ao criar:', error);
       this.loadingService.hideLoading();
     } finally {
