@@ -2,10 +2,10 @@ import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { Product } from '../../../Models/Product';
 import { ProductService } from '../../../Services/ProductService';
-import { AlertService } from '../../../Services/AlertService';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
-  selector: 'app-delete-product',
+  selector: 'app-products/delete-product',
   templateUrl: './delete-product.component.html',
   styleUrls: ['./delete-product.component.css']
 })
@@ -15,7 +15,7 @@ export class DeleteProductComponent implements OnInit {
     public dialogRef: MatDialogRef<DeleteProductComponent>,
     @Inject(MAT_DIALOG_DATA) public data: { product: Product },
     private productService: ProductService,
-    private alertService: AlertService
+    private toastr: ToastrService
     ) { }
 
   ngOnInit(): void {
@@ -29,11 +29,11 @@ export class DeleteProductComponent implements OnInit {
     try {
       if (this.product && this.product.id) {
         await (await this.productService.deleteProduct(this.product.id)).toPromise();
-        this.alertService.success(`Produto ${this.product.name} deletado com sucesso.`);
+        this.toastr.success(`Produto ${this.product.name} deletado com sucesso.`);
         this.dialogRef.close({ deleted: true });
       }
     } catch (error: any) {
-      this.alertService.error(error.message || 'Erro interno da aplicação, tente novamente.');
+      this.toastr.error(error.message || 'Erro interno da aplicação, tente novamente.');
       console.error('Erro ao deletar produto:', error);
     }
   }
