@@ -1,9 +1,10 @@
-import { HttpClient, HttpEvent, HttpHeaders } from '@angular/common/http';
+import { HttpClient, HttpEvent, HttpHeaders, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, catchError, map, switchMap, throwError } from 'rxjs';
 import { environment } from '../../environments/environment';
 import { AuthService } from './AuthService';
 import { SalesRecord } from '../Models/SalesRecord';
+import { PagedResult } from '../Models/PagedResult';
 
 @Injectable({
   providedIn: 'root',
@@ -14,7 +15,25 @@ export class SalesRecordService {
   constructor(private http: HttpClient, private auth: AuthService) { }
 
   getSalesRecords(): Observable<SalesRecord[]> {
-    return this.http.get<SalesRecord[]>(`${environment.apiUrl}/${this.url}/get-salesrecords`);
+    return this.http.get<SalesRecord[]>(`${environment.apiUrl}/${this.url}/get-salesrecords`).pipe(
+      catchError((error: any) => {
+        console.error('Erro HTTP:', error);
+        return throwError(error);
+      })
+    );
+  }
+
+  getSalesRecordsPaginated(page: number = 1, pageSize: number = 10): Observable<PagedResult<SalesRecord>> {
+    let params = new HttpParams()
+      .set('page', page.toString())
+      .set('pageSize', pageSize.toString());
+    return this.http.get<PagedResult<SalesRecord>>(`${environment.apiUrl}/${this.url}/get-salesrecords-paginated`, { params })
+      .pipe(
+        catchError((error: any) => {
+          console.error('Erro HTTP:', error);
+          return throwError(error);
+        })
+      );
   }
 
   async getSalesRecordtById(id: number): Promise<Observable<SalesRecord>> {
