@@ -18,6 +18,27 @@ public class ProductService
                         .Include(x => x.Department)
                         .ToListAsync();
     }
+
+    public async Task<List<Product>> FindAllPaginatedAsync(int pageNumber = 1, int pageSize = 10)
+    {
+        if (pageNumber <= 0 || pageSize <= 0)
+            throw new ArgumentException("Page number and page size must be greater than 0.");
+
+        return await _context.Product
+                              .Include(x => x.Category)
+                              .Include(x => x.Department)
+                              .OrderBy(x => x.Id)
+                              .Skip((pageNumber - 1) * pageSize)
+                              .Take(pageSize)
+                              .ToListAsync();
+    }
+
+    //This is used for counting the number of records for the paginator in FRONTEND.
+    public async Task<int> CountAllAsync()
+    {
+        return await _context.Product.CountAsync();
+    }
+
     public async Task<Product> FindByIdAsync(int id)
     {
         return await _context.Product
@@ -46,13 +67,18 @@ public class ProductService
 
     public async Task<List<Product>> FindByNameAsync(string productName, int? categoryId)
     {
-        var query = _context.Product.Where(p => EF.Functions.Like(p.Name, $"%{productName}%"));
+        var query = _context.Product
+            .Where(p => EF.Functions.Like(p.Name, $"%{productName}%"));
+
+        query = query.Include(p => p.Department) 
+                     .Include(p => p.Category) 
+                     .AsQueryable(); 
 
         if (categoryId.HasValue)
         {
             query = query.Where(p => p.CategoryId == categoryId.Value);
         }
-
         return await query.ToListAsync();
     }
+
 }
